main: initialise 'main.py'
Initialise 'main.py' containing the functions: 'key_press()' and
'key_release(). They take actions when a key is pressed and released
respectively, the former logging the key into a text file. The latter
quits the keylogger when the escape key is pressed.

The advantage of the 'pynput' library is that it can log keystrokes from
any process/thread, something which was hard to achieve (albeit not
impossible) using the Windows API.

Signed-off-by: Shourya Shukla <periperidip@gmail.com>