main: improve key parsing for quotes
The logged keys, when converted to a string to write to a file are
surrounded by quotes, except when the logged string is a single quote,
it is enclosed by double quotes. The keylog is not very reader-friendly
with the quotes.

Therefore, improve the key parsing by taking into account the
surrounding quotation marks and trim them off the logged text
appropriately.

Signed-off-by: Shourya Shukla <periperidip@gmail.com>