helpers: improve space and shift key parsing
pynput logs space and shift presses as 'Key.space' and 'Key.shift'
respectively. In the case of the former, it does not insert the space
but rather logs 'Key.space' and for the latter, it shows the correct
effect when a the shift key is pressed but logs a 'Key.shift' along with
the result of a key press + shift.

Amend the 'write_to_file()' function to account for this and thus give a
more helpful output. While at it, add a newline at the end of the file.

Signed-off-by: Shourya Shukla <periperidip@gmail.com>